requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.18.1
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.6
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wrapt==2.1.2
